# app/libs/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/category/": {
            "get": {
                "description": "Paginate categories using query filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Paginate categories",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "Page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "PageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "tenantId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Category name",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/category_gateway.PaginateOutput"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Create category",
                "parameters": [
                    {
                        "description": "Category data",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/category_gateway.CreateInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created category ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Validation errors",
                        "schema": {
                            "$ref": "#/definitions/middlewares.ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/category/{id}": {
            "get": {
                "description": "Get a category by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Get category by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/category_gateway.GetByIDOutput"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a category by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Delete category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Optional delete filter data",
                        "name": "category",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/category_gateway.DeleteFilter"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update an existing category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Patch category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Category patch data",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/category_gateway.PatchValues"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Validation errors",
                        "schema": {
                            "$ref": "#/definitions/middlewares.ValidationErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/order/": {
            "get": {
                "description": "Paginate orders using query filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Paginate orders",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "Page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "PageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "tenantId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Payment status",
                        "name": "paymentStatus",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fulfillment status",
                        "name": "fulfillmentStatus",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Delivery status",
                        "name": "deliveryStatus",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Created at from (RFC3339)",
                        "name": "createdAtFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Created at to (RFC3339)",
                        "name": "createdAtTo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order_gateway.PaginateOutput"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Create order",
                "parameters": [
                    {
                        "description": "Order data",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/order_case.CreateInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created order ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Validation errors",
                        "schema": {
                            "$ref": "#/definitions/middlewares.ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/order/{id}/payments/confirm": {
            "post": {
                "description": "Confirm payment for a specific order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Confirm order payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payment confirmation data",
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/order_case.ConfirmPaymentInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment confirmed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Validation errors",
                        "schema": {
                            "$ref": "#/definitions/middlewares.ValidationErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errs.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "category_gateway.CreateInput": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tenantId": {
                    "type": "string"
                }
            }
        },
        "category_gateway.DeleteFilter": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "tenantId": {
                    "type": "string"
                }
            }
        },
        "category_gateway.GetByIDOutput": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tenantId": {
                    "type": "string"
                }
            }
        },
        "category_gateway.PaginateData": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "category_gateway.PaginateOutput": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/category_gateway.PaginateData"
                    }
                },
                "maxPages": {
                    "type": "integer"
                }
            }
        },
        "category_gateway.PatchValues": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "errs.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "middlewares.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/validator.ValidationError"
                    }
                }
            }
        },
        "order_case.ConfirmPaymentInput": {
            "type": "object",
            "required": [
                "externalPaymentId",
                "paidAt",
                "provider"
            ],
            "properties": {
                "externalPaymentId": {
                    "type": "string"
                },
                "orderID": {
                    "type": "string"
                },
                "paidAt": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                }
            }
        },
        "order_case.CreateInput": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/order_case.OrderItem"
                    }
                },
                "notes": {
                    "type": "string"
                }
            }
        },
        "order_case.OrderItem": {
            "type": "object",
            "properties": {
                "productId": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "order_gateway.PaginateData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deliveryStatus": {
                    "type": "string"
                },
                "fulfillmentStatus": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "paymentStatus": {
                    "$ref": "#/definitions/payment.Status"
                },
                "tenantId": {
                    "type": "string"
                },
                "totalPrice": {
                    "type": "integer"
                }
            }
        },
        "order_gateway.PaginateOutput": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/order_gateway.PaginateData"
                    }
                },
                "maxPages": {
                    "type": "integer"
                }
            }
        },
        "payment.Status": {
            "type": "string",
            "enum": [
                "pending",
                "authorized",
                "paid",
                "failed",
                "cancelled",
                "refunded"
            ],
            "x-enum-comments": {
                "StatusAuthorized": "auth ok, yet to capture",
                "StatusCancelled": "voided before capture",
                "StatusFailed": "attempt failed",
                "StatusPaid": "captured/settled",
                "StatusPending": "not initiated or waiting user action",
                "StatusRefunded": "refunded after capture"
            },
            "x-enum-descriptions": [
                "not initiated or waiting user action",
                "auth ok, yet to capture",
                "captured/settled",
                "attempt failed",
                "voided before capture",
                "refunded after capture"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusAuthorized",
                "StatusPaid",
                "StatusFailed",
                "StatusCancelled",
                "StatusRefunded"
            ]
        },
...
        "validator.ValidationError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                }
            }
        }
    }
}