# app/libs/http/controllers/category_controller/fiber_create.go
package category_controller

import (
	"net/http"
	"rabi-food-core/libs/database/gateways/category_gateway"
	"rabi-food-core/libs/validator"

	"github.com/gofiber/fiber/v2"
)

// CreateCategory godoc
// @Summary Create category
// @Description Create a new category
// @Tags categories
// @Accept json
// @Produce text/plain
// @Param category body category_gateway.CreateInput true "Category data"
// @Success 201 {string} string "Created category ID"
// @Failure 400 {object} middlewares.ValidationErrorResponse "Validation errors"
// @Failure 500 {string} string "Internal server error"
// @Router /category/ [post]
func (c *CategoryController) Create(ctx *fiber.Ctx) error {
	data := category_gateway.CreateInput{}
	err := ctx.BodyParser(&data)
	if err != nil {
		return ctx.JSON(err)
	}

	err = validator.V.Struct(data)
	if err != nil {
		return err
	}

	id, err := c.usecase.Create(ctx.UserContext(), data)

	if err != nil {
		return err
	}

	return ctx.Status(http.StatusCreated).SendString(id)
}