# app/libs/http/controllers/category_controller/fiber_paginate.go
package category_controller

import (
	"rabi-food-core/libs/database"
	"rabi-food-core/libs/database/gateways/category_gateway"
	"strconv"

	"github.com/gofiber/fiber/v2"
)

// PaginateCategories godoc
// @Summary Paginate categories
// @Description Paginate categories using query filters
// @Tags categories
// @Produce json
// @Param Page query int false "Page number"
// @Param PageSize query int false "Page size"
// @Param tenantId query string false "Tenant ID"
// @Param name query string false "Category name"
// @Success 200 {object} category_gateway.PaginateOutput
// @Failure 500 {string} string "Internal server error"
// @Router /category/ [get]
func (c *CategoryController) Paginate(ctx *fiber.Ctx) error {
	page, err := strconv.Atoi(ctx.Query("Page", "0"))
	if err != nil {
		return err
	}

	pageSize, err := strconv.Atoi(ctx.Query("PageSize", "10"))
	if err != nil {
		return err
	}

	filter := category_gateway.PaginateFilter{}
	err = ctx.QueryParser(&filter)
	if err != nil {
		return err
	}

	paginate := database.PaginateInput{
		Page:     page,
		PageSize: pageSize,
	}

	result, err := c.usecase.Paginate(ctx.UserContext(), filter, paginate)

	if err != nil {
		return err
	}

	return ctx.JSON(result)
}